refactor(proof): usecase 내 dto 패키지 추가
