+++ src/main/java/me/jinjjahalgae/domain/proof/mapper/ProofMapper.java
@@ -6,12 +6,12 @@ import me.jinjjahalgae.domain.feedback.mapper.FeedbackMapper;
 import me.jinjjahalgae.domain.proof.entities.Proof;
 import me.jinjjahalgae.domain.proof.entities.ProofImage;
 import me.jinjjahalgae.domain.proof.enums.ProofStatus;
-import me.jinjjahalgae.domain.proof.usecase.get.await.ProofAwaitResponse;
-import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.await.dto.ProofAwaitResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.dto.ContractorProofListResponse;
 import me.jinjjahalgae.domain.proof.usecase.getlist.common.ProofSimpleResponse;
-import me.jinjjahalgae.domain.proof.usecase.get.detail.ProofDetailResponse;
-import me.jinjjahalgae.domain.proof.usecase.get.recent.ProofRecentResponse;
-import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.SupervisorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.dto.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.recent.dto.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.dto.SupervisorProofListResponse;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
