+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/get/detail/GetProofDetailUseCase.java
@@ -1,5 +1,7 @@
 package me.jinjjahalgae.domain.proof.usecase.get.detail;
 
+import me.jinjjahalgae.domain.proof.usecase.get.detail.dto.ProofDetailResponse;
+
 public interface GetProofDetailUseCase {
 
     ProofDetailResponse execute(Long proofId, Long userId);
