+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/get/recent/GetRecentProofUseCase.java
@@ -1,5 +1,7 @@
 package me.jinjjahalgae.domain.proof.usecase.get.recent;
 
+import me.jinjjahalgae.domain.proof.usecase.get.recent.dto.ProofRecentResponse;
+
 import java.util.List;
 
 public interface GetRecentProofUseCase {
