+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getlist/contractorlist/GetContractorProofListUseCase.java
@@ -1,5 +1,7 @@
 package me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist;
 
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.dto.ContractorProofListResponse;
+
 import java.util.List;
 
 public interface GetContractorProofListUseCase {
