+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getlist/contractorlist/dto/ContractorProofListResponse.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist;
+package me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import me.jinjjahalgae.domain.proof.usecase.getlist.common.ProofSimpleResponse;
