+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getlist/supervisorlist/GetSupervisorProofListUseCase.java
@@ -1,5 +1,7 @@
 package me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist;
 
+import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.dto.SupervisorProofListResponse;
+
 import java.util.List;
 
 public interface GetSupervisorProofListUseCase {
