+++ src/main/java/me/jinjjahalgae/presentation/api/ProofController.java
@@ -6,15 +6,15 @@ import me.jinjjahalgae.domain.proof.usecase.create.common.ProofCreateRequest;
 import me.jinjjahalgae.domain.proof.usecase.create.proof.CreateProofUseCase;
 import me.jinjjahalgae.domain.proof.usecase.create.reproof.CreateReProofUseCase;
 import me.jinjjahalgae.domain.proof.usecase.get.await.GetAwaitProofUseCase;
-import me.jinjjahalgae.domain.proof.usecase.get.await.ProofAwaitResponse;
-import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.await.dto.ProofAwaitResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.dto.ContractorProofListResponse;
 import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.GetContractorProofListUseCase;
 import me.jinjjahalgae.domain.proof.usecase.get.detail.GetProofDetailUseCase;
-import me.jinjjahalgae.domain.proof.usecase.get.detail.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.dto.ProofDetailResponse;
 import me.jinjjahalgae.domain.proof.usecase.get.recent.GetRecentProofUseCase;
-import me.jinjjahalgae.domain.proof.usecase.get.recent.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.recent.dto.ProofRecentResponse;
 import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.GetSupervisorProofListUseCase;
-import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.SupervisorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.dto.SupervisorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
 import me.jinjjahalgae.presentation.api.docs.proof.ProofControllerDocs;
