+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/GetRecentProofsSwaggerResponse.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.presentation.api.docs.proof;
 
-import me.jinjjahalgae.domain.proof.usecase.get.recent.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.recent.dto.ProofRecentResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 
 public class GetRecentProofsSwaggerResponse extends CommonResponse<ProofRecentResponse> {
