+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/ProofControllerDocs.java
@@ -9,11 +9,11 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import me.jinjjahalgae.domain.proof.usecase.create.common.ProofCreateRequest;
-import me.jinjjahalgae.domain.proof.usecase.get.await.ProofAwaitResponse;
-import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.ContractorProofListResponse;
-import me.jinjjahalgae.domain.proof.usecase.get.detail.ProofDetailResponse;
-import me.jinjjahalgae.domain.proof.usecase.get.recent.ProofRecentResponse;
-import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.SupervisorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.await.dto.ProofAwaitResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.dto.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.detail.dto.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.get.recent.dto.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.supervisorlist.dto.SupervisorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
 import org.springframework.web.ErrorResponse;
