+++ src/test/java/me/jinjjahalgae/domain/proof/usecase/getcontractorprooflist/GetContractorProofListUseCaseUnitTest.java
@@ -4,7 +4,7 @@ import me.jinjjahalgae.domain.contract.repository.ContractRepository;
 import me.jinjjahalgae.domain.proof.entities.Proof;
 import me.jinjjahalgae.domain.proof.enums.ProofStatus;
 import me.jinjjahalgae.domain.proof.repository.ProofRepository;
-import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.dto.ContractorProofListResponse;
 import me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist.GetContractorProofListUseCaseImpl;
 import me.jinjjahalgae.domain.proof.util.ProofTestUtil;
 import me.jinjjahalgae.global.exception.AppException;
